Implement #[unix_sigpipe = "sig_ign"] with a noop SIGPIPE handler
Install a noop `SIGPIPE` handler to make `SIGPIPE` ignored in the
current process but automatically set to `SIG_DFL` for child processes,
if the unstable `#[unix_sigpipe = "sig_ign"]` attribute is used (and
only in that case).

Only use this trick for the unstable `#[unix_sigpipe = "sig_ign"]`
attribute and not for the default `SIGPIPE` behavior since it creates
observable side effects. In particular, the handler for `SIGPIPE` can be
observed to not be `SIG_IGN` but instead the address of some function
(namely `_rustc_sigpipe_sigaction_noop()`).