Move `fn is_from_async_await()` from `rustc_middle` to `clippy`
The helper function has only ever been used in `clippy`, and is
currently only used in clippy. Let's move it there. If needed later, it
can be moved to a more central place in `clippy`, or to `rustc_lint`. It
should not be moved back to `rustc_middle` since there is a general
desire to make that crate smaller.