tests: Require `run-fail` ui tests to have an exit code (`SIGABRT` not ok)
Normally a `run-fail` ui test shall not be terminated by a signal like
`SIGABRT`. So begin having that as a hard requirement.

Some of our current tests do terminate by a signal however. Introduce
and use `run-crash` for those tests.

And some tests behave differently on different targets. Introduce and
use `run-fail-or-crash` for those tests. Some sanitizer tests use this
for example, because on some platforms sanitizers will exit with code 1
in the face of memory errors, while on others they exit with abort in
the face of memory errors.