Fix backtraces with `-C panic=abort` on linux; emit unwind tables by default
The linux backtrace unwinder relies on unwind tables to work properly.
The default panic hook prints backtraces. Backtraces with
`-C panic=abort` used to work in Rust 1.22 but broke in Rust 1.23, because
in 1.23 we stopped emitting unwind tables with `-C panic=abort` (see
24cc38e3b00).

In 1.45 (see cda994633ee) a workaround in the form of
`-C force-unwind-tables=yes` was added.

More history

`-C panic=abort` was added in [Rust
1.10](https://blog.rust-lang.org/2016/07/07/Rust-1.10/#what-s-in-1-10-stable)
and the motivation was binary size and compile time. But given how
confusing that behavior has turned out to be, it is better to make
binary size optimzation opt-in with `-C force-unwind-tables=no` rather
than default since the current default breaks backtraces.

Besides, if binary size is a primary concern, there are many other
tricks that can be used that has a higher impact.