tests: basic-[debugger-]stepping.rs: Disable SingleUseConsts MIR pass
To make more of the test pass. It should not be necessary to disable
this pass, but by doing it we can avoid further regressions while we
figure out how to solve this use case properly.

The last two `FIXME(33013)` didn't go away from this change, only the
first six. But that can be investigated separately.