rustc_codegen_llvm: Require opt-level >= 1 for index-based loop
To make debugger stepping intuitive with -Copt-level=0.

This is kind of a revert of bd0aae92dc76d9 except we don't revert it, we
just make it conditional on opt-level. That commit regressed
basic-stepping.rs, but it was not noticed since it did not exist back
then. I have retroactively bisected to find that out.

It seems messy to sprinkle if-cases inside of the
write_operand_repeatedly() so make the whole function conditional.