TEMP COMMIT until PR 147728 is merged
tests: activate misspelled `gdb-check` in `function-arg-initialization.rs`

In 9253e1206e91f5bd7 a bunch of `gdbr-check` (for `rust-gdb`) and
`gdbg-check` (for plain `gdb`) was added. But in two places the author
accidentally wrote `gdbt-check` instead. This commit fixes this typo.

Note that we need to add a `std::hint::black_box()` to avoid

    $4 = <optimized out>

prints on at least `aarch64-gnu-llvm-20-1`.